third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/property-escape-comment-02-expected.txt

CONSOLE ERROR: line 3: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-property.pl?q=%22%20onload=%22alert(2)/' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
